+++ docs/history.rst
@@ -4,6 +4,9 @@ Version History
 Next
 ----
 - Change ``HTTPResponse.links`` to return empty list when ``Link`` header is not present
+- Move ``X-Retry-Attempt`` header insertion into
+  :meth:`~sprockets.mixins.http.HTTPClientMixin._http_req_modify_for_retry`
+  hook.  This is also needed for using the client with OAuth 1 servers.
 
 `2.4.1`_ Nov 30 2020
 --------------------